requirements.txt
tensorflow==2.13.0
flask
numpy
pillow
gdown